Fix issue where + New Note button is offscreen
My son was trying to get your example code to work on the latest XCode / iOS SDK, and we were puzzled by the empty screen that appears by using the code presented here.

It looks like the `Spacer()` is pushing the button offscreen, so it doesn't appear at all. Removing the line fixes the issue.